# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employees API",
        "description": "These are APIs for all Data of Employees in Our Database",
        "contact": {
            "email": "elhirwa3@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/employee": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Display list of All Employees.",
                "operationId": "aec9115c2e1b82bd5139f3ba90d8cd76",
                "responses": {
                    "401": {
                        "description": "UnAuthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UnAuthanticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there is server problem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success|Returns Employees list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Elvin"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "elhirwa3@gmail.com"
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "006"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create an Employee",
                "operationId": "6a4f1e0943433f32bb808e136da78430",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "code": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "post": {
                                                    "type": "string"
                                                },
                                                "identity_number": {
                                                    "type": "string"
                                                },
                                                "age": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Bebe",
                                    "email": "example@content.com",
                                    "code": "101",
                                    "phone": "0784562344",
                                    "post": "Engineer",
                                    "identity_number": "1198580045563328",
                                    "age": "36"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@ab.com"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "404"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "0798888888"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UnAuthanticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there is server problem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "419": {
                        "description": "CSRF Token mismatch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/employee/search/{name}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "operationId": "2b4d3cbb0b536d03579504844b163d3f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "content@gmil.com"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "808"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "089564584"
                                        },
                                        "post": {
                                            "type": "string",
                                            "example": "post"
                                        },
                                        "identity_number": {
                                            "type": "string",
                                            "example": "120859495989598"
                                        },
                                        "age": {
                                            "type": "string",
                                            "example": "40"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UnAuthanticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there is server problem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}